requirements.txt
asgiref==3.8.1
Django==4.2.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
